opencv_requirements.txt
numpy==1.24.4
opencv-contrib-python==4.8.1.78
pillow